models/requirements.txt
django-phonenumber-field==1.1.0
djangorestframework==3.4.6
mysqlclient==1.3.7
django-filter==0.15.2
django-crispy-forms==1.6.0
